requirements.txt
# AI.AS MVP - Requirements
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# AI / ML
numpy>=1.24.0,<2.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers==4.44.0
Pillow>=10.0.0
faiss-cpu==1.8.0

# Data
pandas>=2.0.0
openpyxl>=3.1.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
